Fix Z3 term equality and enable warning-clean tests
Z3Term::compare previously treated matching Z3 hashes as term equality, which allowed distinct ASTs such as integer numerals 9 and -9 to compare equal and collapse in UnorderedTermSet/UnorderedTermMap. Compare Z3 terms and function declarations with Z3_is_eq_ast/Z3_is_eq_func_decl instead, and reject comparisons across different Z3 contexts.

Add a shared parametrized term equality regression that runs against all available integer-capable solver configurations, covering signed numeral construction from both integers and strings as well as unordered term containers. The regression replaces the narrower Z3-only test target.

Enable target-scoped compiler warnings for the core library, solver/test targets, and helper test libraries. Clean the warnings surfaced on the configured Z3 build path, including catch-by-value in tests, signed/unsigned loop indices, initializer order, ignored pipe/write return values, and the accidental std:wstring typo.

Also remove the unnecessary gmock dependency from exec-utils by using direct unordered_set lookup instead of a gmock matcher, allowing configure to work with system GTest installations that do not provide GTest::gmock.

Validation: clang-tidy -p build tests/test-generic-solver.cpp tests/test-str.cpp tests/unit/unit-term.cpp tests/unit/unit-walker.cpp tests/test-term-equality.cpp; cmake --build build --clean-first; ctest --test-dir build -R 'test-term-equality|z3_test' --output-on-failure.

+++ z3/src/z3_sort.cpp
@@ -60,12 +60,12 @@ SortVec Z3Sort::get_domain_sorts() const
 {
   if (is_function)
   {
-    int32_t s_arity = z_func.arity();
+    unsigned int s_arity = z_func.arity();
     SortVec sorts;
     sorts.reserve(s_arity);
     Sort s;
 
-    for (size_t i = 0; i < s_arity; i++)
+    for (unsigned int i = 0; i < s_arity; i++)
     {
       s.reset(new Z3Sort(z_func.domain(i), *ctx));
       sorts.push_back(s);
